style: BodyCondition타입 파일 분할.
